+++ iftttie/services/nest.py
@@ -41,8 +41,8 @@ def yield_updates(data: Any) -> Iterable[Update]:
     devices = data['devices']
 
     yield from yield_devices_updates(data['structures'], 'nest:structure', [
-        ('away', ValueKind.OTHER),
-        ('wwn_security_state', ValueKind.OTHER),
+        ('away', ValueKind.ENUM),
+        ('wwn_security_state', ValueKind.ENUM),
     ])
     yield from yield_devices_updates(devices['cameras'], 'nest:camera', [
         ('is_streaming', ValueKind.BOOLEAN),
@@ -53,7 +53,8 @@ def yield_updates(data: Any) -> Iterable[Update]:
         ('ambient_temperature_c', ValueKind.CELSIUS),
         ('humidity', ValueKind.HUMIDITY),
         ('is_online', ValueKind.BOOLEAN),
-        ('hvac_state', ValueKind.OTHER),
+        ('hvac_state', ValueKind.ENUM),
+        ('target_temperature_c', ValueKind.CELSIUS),
     ])
     yield from yield_devices_updates(devices['smoke_co_alarms'], 'nest:smoke_co_alarm', [
         ('is_online', ValueKind.BOOLEAN),