+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-### `0.2`
+### `latest`
 
 - Improve `Dockerfile` and `Makefile`
+- Improve `Nest` and `Buienradar`
+- Move text constants to templates
+- Introduce new value kinds
 
 ### `0.1`